backend_core/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==42.0.5
Django==5.0.2
django-cors-headers==4.3.1
django-localflavor==4.0
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
phonenumberslite==8.13.32
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python-stdnum==1.20
pytz==2024.1
simplejwt==2.0.1
sqlparse==0.4.4
typing==3.7.4.3
tzdata==2024.1
